Individual Notifications
Added more exception handling and set notifications to show item name and send to the relevent user only.  Checks for invalid urls will send user a push notification saying which listing has an invalid url. If a http request fails, it will recheck the item up to 3 times. If it still cannot successfully scrape the price, the url will be logged to follow up.